+++ Sources/StreamChat/APIClient/Endpoints/ModerationEndpoints.swift
@@ -4,18 +4,6 @@
 
 import Foundation
 
-// MARK: - User muting
-
-extension Endpoint {
-    static func muteUser(_ userId: UserId) -> Endpoint<EmptyResponse> {
-        muteUser(true, with: userId)
-    }
-
-    static func unmuteUser(_ userId: UserId) -> Endpoint<EmptyResponse> {
-        muteUser(false, with: userId)
-    }
-}
-
 // MARK: - User banning
 
 extension Endpoint {
@@ -96,20 +84,6 @@ extension Endpoint {
                 targetUserId: nil,
                 custom: extraData
             )
-        )
-    }
-}
-
-// MARK: - Private
-
-private extension Endpoint {
-    static func muteUser(_ mute: Bool, with userId: UserId) -> Endpoint<EmptyResponse> {
-        .init(
-            path: .muteUser(mute),
-            method: .post,
-            queryItems: nil,
-            requiresConnectionId: false,
-            body: ["target_id": userId]
         )
     }
 }