+++ TestTools/StreamChatTestTools/TestData/DecodableEntity.swift
@@ -13,3 +13,4 @@ extension MessagePayload: DecodableEntity {}
 extension MessageReactionPayload: DecodableEntity {}
 extension UserPayload: DecodableEntity {}
 extension ChannelDetailPayload: DecodableEntity {}
+extension CurrentUserPayload: DecodableEntity {}